+++ plan-service/src/main/java/scdy/planservice/service/PlanDetailService.java
@@ -1,13 +1,15 @@
 package scdy.planservice.service;
 
 import lombok.RequiredArgsConstructor;
+import org.apache.kafka.common.protocol.types.Field;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import scdy.planservice.common.exceptions.NotFoundException;
+import scdy.planservice.common.exceptions.BadRequestException;
 import scdy.planservice.dto.PlanDetailRequestDto;
 import scdy.planservice.dto.PlanDetailResponseDto;
 import scdy.planservice.entity.Plan;
 import scdy.planservice.entity.PlanDetail;
+import scdy.planservice.repository.MemberRepository;
 import scdy.planservice.repository.PlanDetailRepository;
 import scdy.planservice.repository.PlanRepository;
 
@@ -19,6 +21,7 @@ import java.util.List;
 public class PlanDetailService {
     private final PlanDetailRepository planDetailRepository;
     private final PlanRepository planRepository;
+    private final MemberRepository memberRepository;
 
     // 세부 일정 생성
     @Transactional
@@ -48,17 +51,22 @@ public class PlanDetailService {
     }
 
     //일정별 세부 일정 조회
-    public List<PlanDetailResponseDto> readPlanDetailByPlan(Long planId){
+    public List<PlanDetailResponseDto> readPlanDetailByPlan(Long userId, Long planId){
         Plan plan = planRepository.findByIdOrElseThrow(planId);
+        if(!checkAuth(userId, planId)) {
+            throw new BadRequestException("조회 권한이 없습니다.");
+        }
         List<PlanDetail> planDetailList = planDetailRepository.findByPlanId(planId);
-
         return planDetailList.stream().map(PlanDetailResponseDto::from).toList();
     } // repository 쿼리 DSL 적용 필요
 
     // 세부 일정 수정 (순서)
     @Transactional
-    public PlanDetailResponseDto updatePlanDetail(Long planId, Long planDetailId, Integer planDetailtimeLine, Integer planDetailDay){
+    public PlanDetailResponseDto updatePlanDetail(Long userId, Long planDetailId, Integer planDetailtimeLine, Integer planDetailDay){
         PlanDetail planDetail = planDetailRepository.findByPlanDetailIdOrElseThrow(planDetailId);
+        if(!isMember(userId, planDetail.getPlan().getPlanId())){
+            throw new BadRequestException("멤버만 수정할 수 있습니다.");
+        }
         planDetail.updatePlanDetailTimeLine(planDetailtimeLine, planDetailDay);
 
         return PlanDetailResponseDto.from(planDetail);
@@ -67,18 +75,34 @@ public class PlanDetailService {
 
     // 세부 일정 수정 (컨텐츠)
     @Transactional
-    public PlanDetailResponseDto updatePlanDetailContent(Long planDetailId, PlanDetailResponseDto planDetailResponseDto){
+    public PlanDetailResponseDto updatePlanDetailContent(Long userId, Long planDetailId, PlanDetailRequestDto planDetailRequestDto){
         PlanDetail planDetail = planDetailRepository.findByPlanDetailIdOrElseThrow(planDetailId);
-        planDetail.updatePlanDetail(planDetail.getPlanDetailName(), planDetail.getPlanDetailMemo(), planDetail.getPlanDetailTime());
+        if(!isMember(userId, planDetail.getPlan().getPlanId())){
+            throw new BadRequestException("멤버만 수정할 수 있습니다.");
+        }planDetail.updatePlanDetail(planDetailRequestDto.getPlanDetailName(), planDetailRequestDto.getPlanDetailMemo(), planDetailRequestDto.getPlanDetailTime());
         return PlanDetailResponseDto.from(planDetail);
     }
 
     // 세부 일정 삭제
     @Transactional
-    public PlanDetailResponseDto deletePlanDetail(Long planDetailId, Long userId){
+    public PlanDetailResponseDto deletePlanDetail(Long userId, Long planDetailId){
         PlanDetail planDetail = planDetailRepository.findByPlanDetailIdOrElseThrow(planDetailId);
-
+        if(memberRepository.findByUserIdPlanId(userId, planDetail.getPlan().getPlanId()).isEmpty()){
+            throw new BadRequestException("멤버만 삭제할 수 있습니다");
+        }
         planDetailRepository.deleteById(planDetailId);
         return PlanDetailResponseDto.from(planDetail);
+    }
+
+    private boolean checkAuth(Long userId, Long planId){
+        Plan plan = planRepository.findByIdOrElseThrow(planId);
+        if(plan.getIsPublic()){
+            return true;
+        }
+        return memberRepository.findByUserIdPlanId(userId, planId).isPresent();
+    }
+
+    private Boolean isMember(Long userId, Long planId){
+        return memberRepository.findByUserIdPlanId(userId, userId).isPresent();
     }
 }
